big refactor; split unisonoric/bisonoric. tests pass / mypy fails
