+++ tests/test_tune_on_layout.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from pyabc2 import Tune
 
 from concertina_helper.tune_on_layout import TuneOnLayout
-from concertina_helper.layouts import cg_anglo_wheatstone_layout
+from concertina_helper.bisonoric import cg_anglo_wheatstone_layout
 
 tune = Tune(Path('tests/g-major.abc').read_text())
